refactor: move `router.dart` to `routing`
