+++ packages/widgetbook/test/src/routing/router_test.dart
@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 import 'package:widgetbook/src/addons/addon_provider.dart';
 import 'package:widgetbook/src/builder/functions/app_builder.dart';
 import 'package:widgetbook/src/builder/provider/builder_provider.dart';
-import 'package:widgetbook/src/navigation/router.dart';
+import 'package:widgetbook/src/routing/router.dart';
 import 'package:widgetbook/src/repositories/selected_use_case_repository.dart';
 import 'package:widgetbook/widgetbook.dart';
 import 'package:widgetbook_core/widgetbook_core.dart';
